feat(RHINENG-10987): Add status filters
This PR adds filtering on "Status" and "Matched status" for the
Signature Details and System Details tables, respectively. Both sets of
filters apply the same filters to the 2 different tables.

The filters apply both to the systems and signatures that appear in the
table, as well as the matches that appear when expanding the system or
signature.

To test:
Go to the Signatures Page
Click a signature to go to the signature details page
Expand a system on the signature details table
Note the match statuses for the given system
Apply a status/matched status filter and see updated data
Do the same for the Systems Page

Note on how filters should work:

If a signature/system has 1 "Not reviewed" match, 1 "In review" match, and 1 "Resolved" match, then a filter of `IN_REVIEW` would show the said signature/system, but the expanded row would only show the "In review" match.
Likewise, a filter of `TEST` would not show the signature/system at all.