feat(RHINENG-11793): No bell icon when new matches 0
New matches column: Remove the Bell icon when the number is 0 AND make the number gray

Remove the bell icon next to 0
Make the 0 text gray

To test:
- Go to signatures/systems
- Select a signature/system and click

See that a system with more than 0 new matches shows a blue bell icon with the number of new matches. If it has 0 new matches, then the bell icon will not be visible and the 0 will be gray.

In addition, if you expand an affected system or matched signature, the dates matched column should show a blue bell icon with the date text in black.