feat(RHINENG-11792): No bellIcon when older than 30 days
When a match is older than 30 days, the Dates matched cells in the NewMatchesTable should not show the blue notification bell icon, even if the status of that match is still "NOT_REVIEWED".

We will need to wait for the backend api to return a `newMatchCount` to properly update the "New Matches" column counts for affected systems and matched signatures.

To test:
- Go to signatures/systems
- Select a signature/system and click
- Expand an affected system/matched signature

See that dates older than 30 days do not have the bell icon even when they have a status of Not reviewed. For a code review, you'll need to force an older date in the code to get the desired result.