fix(RHINENG-11796): Table overflows page
The CodeEditor in the NewMatchesTable causes the table to extend over the screen on a resize. It's because the CodeEditor is not inside of a Td. We need to find the best way to handle this issue.

To test:
- Select Signatures/Systems
- Select a signature/system and click it
- Expand an affected system/matched signature
- Resize the page to a smaller width

By swapping out the CodeEditor with a CodeBlock, we are able to resolve the overflow issue.